Upgrade tlp-cluster to support building Cassandra with Python 3
CASSANDRA-16396 made Debian packaging rely on Python 3 which broke tlp-cluster's Cassandra build process.
A new docker image was generated that uses Python 3 and fixes this.
CASSANDRA-16480 fixed the debian packages to install on Bionic when created with the cassandra-builds project.
Building trunk now relies on that cassandra-builds project instead of tlp-cluster's internal way of packaging branches.